Add missing upgrade permission for DROP TABLES
The upgrade script for version 1.1 contains the SQL statement "ALTER
TABLE ... RENAME" which requires both CREATE and DROP privileges. If the
database user running the upgrade is missing the DROP privilege, the
upgrade script fails in the middle, but doesn't rollback the changes
already made. This half-upgraded state is problematic. Running
the upgrade script as a transaction would prevent this, but CORAL
currently uses the MyISAM storage engine, which doesn't allow
trasactions. This commit fixes the immediate problem by requiring the
DROP privilege before running the upgrade script, but we should
consider changing from MyISAM to a different engine to provide better
protection during upgrades.

Signed-off-by: Remington Steed <rjs7@calvin.edu>