SECURITY: Enforce redirect to login page
Thanks to Ian Walls for reporting this issue. NOTE: This affects all
modules, not just the Resources module. From his report:

The ajax scripts (ajax_htmldata.php and ajax_processing.php) are not
authenticated. Thus, as a malicious user on the open internet with only
the domain name and path of a CORAL install, I can not only read
usernames and passwords for database platforms (if added by the
institution), but also delete resources from the database and run
arbitrary code.

Examples:
  * Read accounts:
    curl 'http://coraldemo.library.tamu.edu/resources/ajax_htmldata.php?action=getAccountDetails&resourceID=27'
  * Delete a resource:
    curl 'http://coraldemo.library.tamu.edu/resources/ajax_processing.php?action=deleteResource&resourceID=27'
  * Upload a malicious script:
    curl 'http://coraldemo.library.tamu.edu/resources/ajax_processing.php?action=uploadAttachment' -H 'Content-Type: multipart/form-data' -F 'myfile=@/home/user/Desktop/myevilscript.php'

This commit simply causes the script to exit after attempting to forward
to the login page, to prevent any further data from being sent to
disobedient browsers (such as curl). A patch file is also included.

Signed-off-by: Remington Steed <rjs7@calvin.edu>